+++ client/src/screens/about/About.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
 
 export const About = () => {
-    return <div>About Screen</div>
+    return <div className="container max-w-screen-md mx-auto flex flex-col pt-16">About Screen</div>
 }
