+++ client/src/screens/splash/Splash.tsx
@@ -22,7 +22,7 @@ export const Splash = () => {
     const [text, setText] = useState('')
 
     return (
-        <>
+        <div className="container max-w-screen-md mx-auto flex flex-col pt-16">
             <H1>Agon Dice Roller</H1>
             <p className="text-base font-normal pb-6">This is a fan-made dice-rolling app for AGON. AGON is an action-packed roleplaying game about epic Heroes who face trials from the Gods in an ancient world of myth and legend. Learn more about it, and the Paragon system, at <A href="http://agon-rpg.com">agon-rpg.com</A></p>
             <H3>Start a New Session</H3>
@@ -44,6 +44,6 @@ export const Splash = () => {
             <p className="text-base pb-3">This app is a lightweight dicerolling app, focused on resolving and narrating AGON Contests. It is not a campaign or character tracker - you'll still need to manage that on your own.</p>
             <p className="text-base pb-3">This app was created by <A href="">@gareth</A> and <A href="">@sporgory</A> with the help of the <A href="">AGON fan Discord community</A>.</p>
             <p className="text-base pb-3 mb-64">If you’d like to give feedback or get involved, check out the <A href="https://github.com/gareththegeek/rollagon">GitHub repo</A>.</p>
-        </>
+        </div>
     )
 }
